==> 1. Download and collect data.ipynb <==
## Download the data

#!wget https://www2.informatik.uni-hamburg.de/nats/pub/SWC/SWC_Dutch_no_audio.tar.xz -O data/raw/nl_wiki_text.tar.xz
!tar -xf data/raw/nl_wiki_text.tar.xz -C data/raw/

import numpy as np # linear algebra
import pandas as pd # data processing, CSV file I/O (e.g. pd.read_csv)
import os

text_path = []

for dirname, _, filenames in os.walk('./data/raw/dutch'):
    for filename in filenames:
        if filename == 'audiometa.txt':
            interested_file = os.path.join(dirname, filename)
            text_path.append(interested_file)

print(f"There are {len(text_path)} files with interesting documents.")

text_list = []
for file in text_path:
    with open(file, 'r') as myfile:
        data = myfile.read()
        text_list.append(data)
        

df = pd.DataFrame({'text' : text_list})

df.to_pickle("./data/processed/df_wiki_text.pickle")
