==> 3. Create the TSNE representation.ipynb <==
import pandas as pd
from pprint import pprint
import numpy as np
from sklearn.manifold import TSNE
import matplotlib.pyplot as plt
import plotly.express as px
import re
from spacy import displacy
from spacy.symbols import NOUN, DET, ADJ


df_processed = pd.read_pickle('./data/dumps/woorden_met_hetofde.pickle')

df_processed.shape

df_processed.head()

tokens = df_processed.woord_vec.tolist()
de_of_het_label = df_processed.det.tolist()
annotation_label =    df_processed.woord.tolist()

tsne_model = TSNE(perplexity=50, n_jobs=3, n_components=2, n_iter=5000, random_state=23)
new_values = tsne_model.fit_transform(tokens)

x = new_values[:,0]
y = new_values[:,1]
df_processed['2d_tsnse_x'] = x
df_processed['2d_tsnse_y'] = y
df_processed.head()

df_processed.to_pickle('./data/dumps/woorden_met_hetofde_tsne_2d_perplexity_50.pickle')
